fix delete cs

+++ functions/Customer-service/delete_cs.php
@@ -19,17 +19,17 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
         // First, delete all pivot references to the instructor
         $queryDelPivot = "DELETE FROM branch_instructor WHERE instructor_id = ?";
         $DelPivotstmt = $pdo->prepare($queryDelPivot);
-        $DelPivotstmt->execute([$cs_id]);
+        $resultDelPivot = $DelPivotstmt->execute([$cs_id]);
 
         // then delete from instructor
-        $query = "DELETE FROM instructors WHERE id = ? AND role IN ('cs', 'cs-admin')";
+        $query = "DELETE FROM instructors WHERE id = ? AND role IN ('cs', 'cs-admin' , 'owner')";
         $stmt = $pdo->prepare($query);
-        $stmt->execute([$cs_id]);
+        $result = $stmt->execute([$cs_id]);
 
         // Commit transaction if both queries succeeded
         $pdo->commit();
 
-        if ($stmt->rowCount() > 0) {
+        if ($resultDelPivot && $result) {
             echo json_encode(['status' => 'success', 'message' => 'CS Agent Deleted successfully']);
         } else {
             echo json_encode(['status' => 'error', 'message' => 'CS Agent not found or could not be deleted']);
